requirements.txt
numpy
tensorflow-gpu
sklearn
opencv-python
pandas
matplotlib
scipy
scikit-image